refactor(Dialog): useEnvironmentの移動に追従
