+++ packages/smarthr-ui/src/components/Dialog/DialogHeader.tsx
@@ -3,7 +3,7 @@
 import { type FC, type PropsWithChildren, useMemo } from 'react'
 import { tv } from 'tailwind-variants'
 
-import { useEnvironment } from '../../hooks/useEnvironment'
+import { useEnvironment } from '../../hooks/client/useEnvironment'
 import { Cluster } from '../Layout'
 
 const classNameGenerator = tv({
