+++ packages/smarthr-ui/src/components/Dialog/ModelessDialog/ModelessDialog.tsx
@@ -19,8 +19,8 @@ import Draggable, { type DraggableBounds } from 'react-draggable'
 import { type VariantProps, tv } from 'tailwind-variants'
 
 import { useAnimationFrame } from '../../../hooks/client/useAnimationFrame'
+import { useEnvironment } from '../../../hooks/client/useEnvironment'
 import { useMergeRefs } from '../../../hooks/client/useMergeRefs'
-import { useEnvironment } from '../../../hooks/useEnvironment'
 import { useHandleEscape } from '../../../hooks/useHandleEscape'
 import { useLatest } from '../../../hooks/useLatest'
 import { Localizer, useIntl } from '../../../intl'
